# src/main/java/dev/kuchishkin/dto/UserNotificationDto.java
package dev.kuchishkin.dto;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@JsonInclude(JsonInclude.Include.NON_NULL)
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class UserNotificationDto {

    private Long id;

    private Long notificationEventId;

    private Long userId;

    private boolean isRead;
}